requirements.txt
flask
selenium
requests
psutil
pyyaml 